backend/requirements.txt
fastapi
uvicorn[standard]
jinja2
httpx
apscheduler
python-dotenv
supabase
pytest
pytest-asyncio